Refactor loadout drawer a bit
