+++ src/app/loadout-drawer/LoadoutDrawerContents.tsx
@@ -73,7 +73,6 @@ export default function LoadoutDrawerContents(
     buckets,
     items,
     stores,
-    itemSortOrder,
     equip,
     remove,
     add,
@@ -85,7 +84,6 @@ export default function LoadoutDrawerContents(
     buckets: InventoryBuckets;
     stores: DimStore[];
     items: DimItem[];
-    itemSortOrder: string[];
     equip(item: DimItem, e: React.MouseEvent): void;
     remove(item: DimItem, e: React.MouseEvent): void;
     add(item: DimItem, e?: MouseEvent, equip?: boolean): void;
@@ -147,7 +145,6 @@ export default function LoadoutDrawerContents(
             bucket={bucket}
             loadoutItems={loadout.items}
             items={itemsByBucket[bucket.hash] || []}
-            itemSortOrder={itemSortOrder}
             pickLoadoutItem={(bucket) => pickLoadoutItem(loadout, bucket, add)}
             equip={equip}
             remove={remove}
